secrets: codex round 15
- move field_summary and its source policy into secrets/summary, so
  errors.py and errors.ts hold only the exception type
