+++ python/mirage/secrets/errors.py
@@ -12,41 +12,6 @@
 # limitations under the License.
 # ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
 
-from collections.abc import Mapping
-
-# Past this many, what came back is not a secret's shape: the `env`
-# source answers with the whole process environment, and reciting a
-# host's variable names back to the agent is neither a useful hint nor
-# ours to print.
-MAX_LISTED_FIELDS = 12
-
-# Sources whose fields are the host's shape rather than a secret's,
-# and are never named back however few of them there are: a hardened
-# container starts from `env -i` plus a handful of credentials, so a
-# count threshold alone would recite exactly the environment worth
-# hiding.
-OPAQUE_FIELD_SOURCES = frozenset({"env"})
-
-
-def field_summary(fields: Mapping[str, str], source: str) -> str:
-    """How a refusal names the fields the secret did carry.
-
-    Lives beside the error it words because both the config plane and
-    the env plane raise it, from different packages.
-
-    Args:
-        fields (Mapping[str, str]): the fetched secret's fields.
-        source (str): the source they came from.
-
-    Returns:
-        str: what follows "has" in the message -- the labels for a
-            secret of ordinary size, a bare count for the process
-            environment or for anything big enough to be one.
-    """
-    if source in OPAQUE_FIELD_SOURCES or len(fields) > MAX_LISTED_FIELDS:
-        return f"{len(fields)} fields"
-    return "{" + ", ".join(sorted(fields)) + "}"
-
 
 class SecretsError(Exception):
     """A secrets source could not answer, or was addressed wrongly.
