+++ python/mirage/workspace/workspace/fill.py
@@ -21,8 +21,9 @@ from mirage.commands.cli.walk import invoked_env_names, supplied_env_names
 from mirage.runtime.base import Runtime
 from mirage.runtime.routing import RouteDecision
 from mirage.runtime.table import VFSRuntime
-from mirage.secrets.errors import SecretsError, field_summary
+from mirage.secrets.errors import SecretsError
 from mirage.secrets.registry import fetch_secret
+from mirage.secrets.summary import field_summary
 from mirage.secrets.types import ResolvedSource
 from mirage.shell.constants import SHOPT_DEFAULTS
 from mirage.shell.parse import (arith_reads, assignment_values,
