+++ typescript/packages/core/src/secrets/errors.ts
@@ -12,8 +12,6 @@
 // limitations under the License.
 // ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
 
-import { compareCodePoints } from '../utils/sort.ts'
-
 /**
  * A secrets source could not answer, or was addressed wrongly.
  *
@@ -28,33 +26,4 @@ export class SecretsError extends Error {
     super(message, options)
     this.name = 'SecretsError'
   }
-}
-
-// Past this many, what came back is not a secret's shape, and
-// reciting a host's names back to the agent is neither a useful hint
-// nor ours to print.
-export const MAX_LISTED_FIELDS = 12
-
-// Sources whose fields are the host's shape rather than a secret's,
-// and are never named back however few of them there are: a hardened
-// container starts from `env -i` plus a handful of credentials, so a
-// count threshold alone would recite exactly the environment worth
-// hiding.
-export const OPAQUE_FIELD_SOURCES: ReadonlySet<string> = new Set(['env'])
-
-/**
- * How a refusal names the fields the secret did carry.
- *
- * Lives beside the error it words because both the config plane and
- * the env plane raise it, from different packages. Returns what
- * follows "has" in the message: the labels for a secret of ordinary
- * size, a bare count for the process environment or for anything big
- * enough to be one.
- */
-export function fieldSummary(fields: Readonly<Record<string, string>>, source: string): string {
-  const names = Object.keys(fields)
-  if (OPAQUE_FIELD_SOURCES.has(source) || names.length > MAX_LISTED_FIELDS) {
-    return `${String(names.length)} fields`
-  }
-  return `{${names.sort(compareCodePoints).join(', ')}}`
 }
