+++ typescript/packages/core/src/workspace/workspace/fill.ts
@@ -16,7 +16,8 @@ import { invokedEnvNames, suppliedEnvNames } from '../../commands/cli/walk.ts'
 import type { Runtime } from '../../runtime/base.ts'
 import type { RouteDecision } from '../../runtime/routing/index.ts'
 import { VFSRuntime } from '../../runtime/table.ts'
-import { SecretsError, fieldSummary } from '../../secrets/errors.ts'
+import { SecretsError } from '../../secrets/errors.ts'
+import { fieldSummary } from '../../secrets/summary.ts'
 import { fetchSecret } from '../../secrets/registry.ts'
 import type { ResolvedSource } from '../../secrets/types.ts'
 import { SHOPT_DEFAULTS } from '../../shell/constants.ts'
